requirements.txt
#tensorflow
tensorflow-gpu
numpy
matplotlib
scipy
